support for release 1.7

+++ webui/models.py
@@ -1,3 +1,6 @@
+__license__= "AGPL-3.0"
+__author__ = "Sucial https://github.com/SUC-DriverOld"
+
 import os
 import shutil
 import requests
@@ -184,6 +187,8 @@ def install_unmsst_model(unmsst_model, unmsst_config, unmodel_class, unmodel_typ
             return i18n("请上传'.yaml'格式的配置文件")
         if not unmsst_model.endswith((".ckpt", ".chpt", ".th")):
             return i18n("请上传'ckpt', 'chpt', 'th'格式的模型文件")
+        if unmodel_class == "" or unmodel_type == "":
+            return i18n("请输入正确的模型类别和模型类型")
 
         if model_name in load_msst_model():
             os.remove(os.path.join(MODEL_FOLDER, unmodel_class, model_name))
@@ -219,6 +224,8 @@ def install_unvr_model(unvr_model, unvr_primary_stem, unvr_secondary_stem, model
         model_name = os.path.basename(unvr_model)
         model_map[model_name] = {}
 
+        if model_param == "":
+            return i18n("请输入选择模型参数")
         if model_param == i18n("上传参数") and not upload_param.endswith(".json"):
             return i18n("请上传'.json'格式的参数文件")
         if not unvr_model.endswith(".pth"):
